put the 2 new rules in WithFields.scala
